Initial commit
mlir-hail is an experimental compiler prototyping a next generation
using the MLIR project

At current only a single CMakeLists.txt is present making sure sanity of
the build environment is used. If, like me, you're installing MLIR
separately, the best way to tell cmake the location of the MLIR
installation is through setting the MLIR_DIR environment variable, like
so:

    $ mkdir build
    $ cd build
    $ MLIR_DIR=/path/to/lib/cmake/mlir cmake .. -G Ninja

If the cmake command executes, then the environment should be reasonably
sane.