Add verification to consume_opt
Add DeclareOpInterfaceMethods, and SingleBlockImplicitTerminator to
ConsumeOptOp. This enables us to use MLIR's inbuilt control flow graph
checks, as well as automatically adding a yield if one isn't present.
MLIR's CFG checks will check if the types of the yields are the same.

Add a static verify method that calls RegionBranchOpInterface::verify
and also checks that the presentRegion's block has the same input types
as the optional type. It is a future TODO to make OptionalType properly
variadic in type parameters.