use UnrealizedConversionCastOp in lowering
Use the builtin op `unrealized_conversion_cast` instead of custom pack/unpack ops. This will help avoid boilerplate in large scale lowering. Also take advantage of the fold operations defined on `unrealized_conversion_cast` to reduce the number of casts inserted by the conversion pass. It still inserts some unnecessary casts, but that seems unavoidable with the current state of support for 1-to-N type conversion in the dialect conversion infrastructure.