Add infrastructure for building this project in Hail Batch
Includes:
- Dockerfiles for both debian and alpine linux.
- A hail batch script that builds LLVM in both debian and alpine.
- Reduction of cmake_minimum_required to 3.18 since that's what debian
  ships

Notes:
In the interest of having a consistent compiler in any eventual
image/CI, I also build clang and lld in the hail-batch script.

We manually install llvm-lit against system python as well. Since it is
not installed as part of any of llvm's install targets.

Furthermore, the docker images and batch script use [sccache]
a distributed compilation cache (like ccache but it supports cloud
object storage). So it's possible that an eventual ci pipeline would be
able to 'build' LLVM on every PR.

Using sccache reduced build times from about an hour to only a few
minutes when I tested it, and that included compressing the resulting
artifacts.

[sccache]: https://github.com/mozilla/sccache