Update tests to look for 'lit' and substitute FileCheck's full path
[The lit library] is a python package that can be installed from PyPI or
user's package managers. This is the same package as what `llvm-lit`
calls, and an examination of `llvm-lit` shows that it searches it's
source tree for the `lit` package and runs it.

Furthermore, FileCheck is often not directly on user's PATH so we add
a substitution to lit.site.cfg.py.in to use the full path.